+++ vc-projects/Blip_n_Blop_3/menus/menu_list.h
@@ -18,7 +18,7 @@ class MenuList {
 
     void MoveDown() {
         focused_ += 1;
-        if (focused_ >= items_.size()) {
+        if ((size_t)focused_ >= items_.size()) {
             focused_ = 0;
         }
     }
